+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## Unreleased
+## Version 0.8.5 (2015-08-12)
 
  - Added support for cubemaps and cubemap arrays. Not all operations are available yet.
  - `DrawParameters` no longer implements `Copy`.